smart_drive/requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Limiter>=3.5.0
Flask-Mail>=0.10.0
pymongo>=4.7.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
Pillow>=10.4.0
WTForms>=3.1.0
email-validator>=2.0.0
itsdangerous>=2.1.0
bleach>=6.1.0
